// src/services/-nav-menus-post.schemas.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Shopping Platform API
 * Production-grade e-commerce backend
 * OpenAPI spec version: 1.0
 */
export type DtoUpsertNavMenuRequestType = typeof DtoUpsertNavMenuRequestType[keyof typeof DtoUpsertNavMenuRequestType];


export const DtoUpsertNavMenuRequestType = {
  mega: 'mega',
  link: 'link',
} as const;

export interface DtoLink {
  href?: string;
  title?: string;
}

export interface DtoColumn {
  links?: DtoLink[];
  title?: string;
}

export interface DtoFeaturedItem {
  badge?: string;
  description?: string;
  href?: string;
  image?: string;
  title?: string;
}

export interface DtoViewAll {
  href?: string;
  label?: string;
}

export interface DtoUpsertNavMenuRequest {
  badge?: string;
  columns?: DtoColumn[];
  featured?: DtoFeaturedItem[];
  href?: string;
  label: string;
  order?: number;
  type: DtoUpsertNavMenuRequestType;
  viewAll?: DtoViewAll;
}

export interface UtilsResponse {
  code?: number;
  error?: string;
  message?: string;
  success?: boolean;
}

export interface DtoNavItemResponse {
  badge?: string;
  columns?: DtoColumn[];
  featured?: DtoFeaturedItem[];
  href?: string;
  id?: number;
  label?: string;
  order?: number;
  type?: string;
  viewAll?: DtoViewAll;
}

export type PostNavMenus201 = UtilsResponse & {
  data?: DtoNavItemResponse;
};